remove redundant code from both classes files
